+++ src/main/resources/application-default.properties
@@ -53,3 +53,9 @@ top.level.column.id.report=3
 app.home.banner.count=6
 
 spring.thymeleaf.cache=false
+
+
+spring.tomcat.https.keystoreAlias=tomcatjks
+spring.tomcat.https.keystore=d://tomcat.jks
+spring.tomcat.https.keystorePass=lirong
+spring.tomcat.https.port=8443
